Update README.md
ABOUT COMPANY

CODSOFT is a vibrant and diverse community that brings together
individuals with similar objectives and ultimate goals. Our main focus
is on creating opportunities that span various areas, including
leadership development, learning, student engagement, and fostering
shared interests.

We believe in the power of leadership and its ability to drive positive
change. That's why we provide platforms and resources for our community
members to develop their leadership skills. Through mentorship programs,
workshops, and collaborative projects, we empower individuals to take on
leadership roles and make a difference in their respective fields.